% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_divide.R
\name{str_divide}
\alias{str_divide}
\title{str_divide}
\usage{
str_divide(str, by = 2, collapse = "\\n")
}
\arguments{
\item{str}{Cadena a ajustar}

\item{by}{Cantidad de veces a dividir la cadena}

\item{collapse}{NA para hacer unsimple split, sino el caracter para dividir
la cadena, por ejemplo `n`}
}
\value{
La cadena ajustada con saltos de línea o un vector de caracteres
}
\description{
Ajuste inteligente del tamaño de una cadena
}
\details{
Agrega saltos de línea, a una cadena, en los espacios en blanco de ésta,
para ajustar el un determinado ancho. Ideal para ajustar los
labels de las gráficas, cuando son muy largos, en más de una línea y
así ahorrar espacio. Esta función es "inteligente" en el sentido que divide
una cadena un numero de veces pero equilibrando los tamaños de las
subcadengas generadas
}
\examples{
str_divide("JUNTOS por el CAMBIO", 2)
str_divide("JUNTOS por el CAMBIO", 1)

}
