% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_label.R
\name{str_label}
\alias{str_label}
\title{**str_label**}
\usage{
str_label(str, width = 5)
}
\arguments{
\item{str}{Cadena a ajustar}

\item{width}{Ancho máximo en caracteres}
}
\value{
La cadena ajustada con saltos de línea
}
\description{
Agrega saltos de línea, a una cadena, en los espacios en blanco de ésta,
para #' ajustar el un determinado ancho. Ideal para ajustar los
__labels__ de las gráficas, cuando son muy largos, en más de una línea y
así ahorrar espacio.
}
\examples{
str_label("Etiqueta demasiado larga", 10)

}
