% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_elegante.R
\name{theme_elegante}
\alias{theme_elegante}
\title{Elegante theme}
\usage{
theme_elegante(base_size = 10, base_family = "Raleway")
}
\arguments{
\item{base_size}{Change the font base size. Default to 10}

\item{base_family}{Change the font family. Defaults to standard Raleway}
}
\value{
ggplot theme
}
\description{
Elegante theme
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(x = mpg*0.43, y = wt*0.4535924, colour = factor(cyl))) +
  geom_point(size = 2) +
  labs(title = "Car weight vs efficiency",
       subtitle = "Using sensible metrics",
       x = "Efficiency (km/l)",
       y = "Weight (1000 kg)",
       colour = "Cylinders") +
  theme_elegante()

}
\seealso{
\code{\link{theme_elegante}}
}
